// frontend/src/components/common/Card.d.ts
import React, { ReactNode } from 'react';
interface CardProps {
    children: ReactNode;
    className?: string;
    title?: string;
    subtitle?: string;
    action?: ReactNode;
    actions?: ReactNode;
}
export declare const Card: React.FC<CardProps>;
export {};
//# sourceMappingURL=Card.d.ts.map